pdf tests moved to pdf_processing folder
