+++ tests/unit/pdf_processing/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for PDF processing functionality."""
