+++ tests/unit/pdf_processing/pdf_text_extraction_test_functions.py
